simplify snapshot scheduling: extract error helper, inline thin wrappers, remove dead code
- Extract failScheduleRun helper to deduplicate the "record error + save
  schedule" pattern repeated 4 times in runSnapshotScheduleForInstanceLocked
- Inline countScheduledSnapshots (single-use len() wrapper) at call site
- Inline validateSetSnapshotScheduleRequest (single-use wrapper) at call site
- Merge snapshot_schedule_types.go contents into snapshot_schedule.go
- Remove dead truncation/fallback code in BuildSnapshotName that could never
  execute because ValidateSetRequest already rejects invalid prefixes upstream
